Further C++11 updates and architecture improvements
Remove requirement for callers to apply filters to newly opened files, adding
a parameter to Archive::open() instead, which controls whether filters are
applied or not to each newly opened file.

Make Archive::open return a unique_ptr to the stream instead of a shared_ptr.

Make FilterType take a unique_ptr so that it takes ownership of the underlying
stream.

Replace more variable declarations with auto.

Make FATArchive return an archfile instance instead, which means it no longer
needs to keep track of open streams and update them when files change.  Now it
can just update the FATEntry values and all open files/substreams will see the
change automatically.
This also fixes the potentially cyclic reference between the FATArchive holding
a reference to each open stream, which needed a custom truncate function which
held a reference to the FATArchive so Archive::resize() could be called.  Now
each archfile stream holds one reference to the FATArchive so Archive::resize()
can be called, and that's it.